Fix race condition
It's possible for the SMTP client to be ready to send the email
even before Connect() returns.  Also, GetReadyOrBrokenFuture
always creates a new promise, so if the condition occurred before
this call, the new promise is never set.

So do not wait until Connect() to get the ready/broken promise future.